Eliminate minor bug in interpolatorND
This fix needed an if-block added when the interpolator is queried at *exactly* the grid maximum along one or more axes.  In this case, the wrong value was returned because the interpolation assumes the reference point is at the lower corner of the hypercube containing the query point.  The if-block is only activated if a point contains exactly the grid maximum point.